+++ app/src/main/java/woowacourse/payments/ui/cardregister/component/PasswordInputField.kt
@@ -1,0 +1,37 @@
+package woowacourse.payments.ui.cardregister.component
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import woowacourse.payments.R
+
+@Composable
+fun PasswordInputField(
+    text: String,
+    onValueChange: (String) -> Unit,
+) {
+    val maxLength = 4
+
+    OutlinedTextField(
+        modifier = Modifier.fillMaxWidth(0.5f),
+        value = text,
+        onValueChange = { newText ->
+            val filteredText = newText.filter { it.isDigit() }
+            if (filteredText.length <= maxLength) onValueChange(filteredText)
+        },
+        label = { Text(stringResource(R.string.password_label)) },
+        placeholder = {
+            Text(
+                text = stringResource(R.string.password_place_holder),
+                color = Color.Gray,
+            )
+        },
+        visualTransformation = PasswordVisualTransformation(),
+        singleLine = true,
+    )
+}
