+++ app/src/main/java/woowacourse/payments/ui/cards/component/SelectBankBottomSheet.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.component
+package woowacourse.payments.ui.cards.component
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -13,15 +13,16 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import woowacourse.payments.ui.cardregister.SelectCardButtons
+import woowacourse.payments.ui.cardregister.BankViewType
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun BottomSheet(
+fun SelectBankBottomSheet(
     modifier: Modifier = Modifier,
     shape: Shape,
     containerColor: Color = Color.White,
-    content: @Composable () -> Unit,
+    showBottomSheet: (Boolean) -> Unit,
+    onCardAddClick: (BankViewType) -> Unit,
     show: Boolean,
     onDismiss: () -> Unit,
 ) {
@@ -30,7 +31,21 @@ fun BottomSheet(
             modifier = modifier,
             shape = shape,
             containerColor = containerColor,
-            content = { content() },
+            content = {
+                Box(
+                    modifier =
+                        Modifier
+                            .fillMaxWidth()
+                            .height(342.dp),
+                    contentAlignment = Alignment.Center,
+                    content = {
+                        SelectCardButtons(onBankClick = { bankType ->
+                            showBottomSheet(false)
+                            onCardAddClick(bankType)
+                        })
+                    },
+                )
+            },
             onDismissRequest = onDismiss,
         )
     }
@@ -39,20 +54,12 @@ fun BottomSheet(
 @Preview
 @Composable
 private fun ShowBottomSheetPreview() {
-    BottomSheet(
+    SelectBankBottomSheet(
         modifier = Modifier.height(342.dp),
         shape = RoundedCornerShape(topStart = 28.dp, topEnd = 28.dp),
-        content = {
-            Box(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .height(342.dp),
-                contentAlignment = Alignment.Center,
-                content = { SelectCardButtons(onBankClick = {}) },
-            )
-        },
         show = true,
         onDismiss = {},
+        showBottomSheet = {},
+        onCardAddClick = {},
     )
 }
